major refactor, include multi-verifier
